pyproject.toml
[tool.poetry]
name = "casalioy"
version = "1.6.0"
description = "Your local langchain toolkit"
authors = ["su77ungr"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">3.9.7,<3.12"
pdfminer-six = "20221105"
python-dotenv = "^1.0.0"
pathlib = "^1.0.1"
sentencepiece = "^0.1.99"  # For convert.py

[tool.poetry.group.GUI.dependencies]  # for the GUI
streamlit = { git = "https://github.com/hippalectryon-0/streamlit.git", subdirectory = "lib", "rev" = "0b7fb1c" }  # waiting for >1.22.0 so that we can use protobuf>=4
streamlit-chat = "^0.0.2.2"
streamlit-extras = "^0.2.7"

[tool.poetry.group.LLM.dependencies]
langchain = "^0.0.168"
pygpt4all = "^1.1.0"
qdrant-client = "^1.1.7"
unstructured = "^0.6.6"  # Handle ingestion file formats
docx2txt="^0.8"  # Handle docx ingestion file formats
llama-cpp-python = "<=0.1.49"  # 0.1.50 raises an AssertionError / NameError for some reason
sentence_transformers = "^2.2.2"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.1"
isort = "5.12.0"
black = "23.3.00"
pycln = "2.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# pre-commit config
[tool.black]
line-length = 160

[tool.pycln]
all = true

[tool.isort]
line_length = 160
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
split_on_trailing_comma = true
